+++ lakefs_provider/operators/merge_operator.py
@@ -4,9 +4,11 @@ from airflow.models import BaseOperator
 from airflow.utils.decorators import apply_defaults
 
 from lakefs_provider.hooks.lakefs_hook import LakeFSHook
+from lakefs_provider.links.lakefs_link import LakeFSLink
+from lakefs_provider.operators.with_metadata_operator import WithLakeFSMetadataOperator
 
 
-class LakeFSMergeOperator(BaseOperator):
+class LakeFSMergeOperator(WithLakeFSMetadataOperator):
     """
     Merge source branch to destination branch
 
@@ -35,6 +37,8 @@ class LakeFSMergeOperator(BaseOperator):
     template_ext = ()
     ui_color = '#f4a460'
 
+    operator_extra_links = [LakeFSLink()]
+
     @apply_defaults
     def __init__(self, lakefs_conn_id: str, repo: str, source_ref: str, destination_branch: str, msg: str, metadata: Dict[str, str] = None, **kwargs: Any) -> None:
         super().__init__(**kwargs)
@@ -51,7 +55,16 @@ class LakeFSMergeOperator(BaseOperator):
         self.log.info("Merging to lakeFS branch '%s' in repo '%s' from source ref '%s'",
                       self.destination_branch, self.repo, self.source_ref)
 
-        self.metadata.__setitem__("airflow_task_id", self.task_id)
+        self.metadata["airflow_task_id"] = self.task_id
+
+        self.enrich_metadata(context)
+
         ref = hook.merge(self.repo, self.source_ref, self.destination_branch, self.msg, self.metadata)
+
+        LakeFSLink.persist(context,
+                           task_instance=self,
+                           lakefs_base_url=hook.get_base_url(),
+                           repo=self.repo,
+                           commit_digest=ref)
 
         return ref